Updated set date and started on google calendar inplemntation
Change a greater than to less in onSelectedDayChanged for the calendar so that only dates in the future can be selected as valid event dates.

Modified TimePickerFragement to include the trailing 0 when time was set to a minute value starting with 0 i.e "14:5" -> "14:05".

Changed the constructor of time picker dialog to be a 24 hour clock.

Started to add Google calendar support, created the key and client it + secret and is now updating the gradle so that the api can be used.