Base notifications are functional
Base notifications are now implemented in this Application! Right now it depends completely on the SQLite database we have set up, the next change would be to have it depend on the spring.io database.

Also currently nothing occurs when the notification is pressed, the plan is to have an "Accept" and "Deny" button which will have differing results when pressed.