Finished all Database integration with Notifications
All that needs to be done is some fine tuning of the asynchronous background tasks and proper documentation of the code.
Finally I'll have to reorganize the firebase function to send as data instead of a notification as then I can handle the notifications even when the application is killed.